+++ pkg/models/program-breaks-light-model.go
@@ -92,7 +92,7 @@ func (cfg *ProgramBreaksLightConfiguration) InitJob() func() {
 		var cnl []Channels
 		months := map[int64][]time.Time{}
 		badgerBudgets := storage.Open(DbBudgets)
-		err = badgerBudgets.Find(&budgets, badgerhold.Where("Month").Ge(-1))
+		err = badgerBudgets.Find(&budgets, badgerhold.Where("Month").Ge(int64(-1)))
 		if err != nil {
 			fmt.Printf("Q:%s - err:%s", qName, err.Error())
 			return
